# swagger.yaml
openapi: 3.0.0
paths:
...
  /profiles:
    post:
      operationId: ProfilesController_create
      summary: Создать новый профиль
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileDto'
      responses:
        '201':
          description: Профиль успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '403':
          description: Отказ в доступе
        '422':
          description: Неверный запрос
      tags: &ref_1
        - profiles
      security: &ref_2
        - bearer: []
    get:
      operationId: ProfilesController_findAll
      summary: Получить все профили
      parameters: []
      responses:
        '200':
          description: Профили успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Profile'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_1
      security: *ref_2
  /profiles/{id}:
    get:
      operationId: ProfilesController_findOne
      summary: Получить профиль по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор профиля
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Профиль успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_1
      security: *ref_2
    patch:
      operationId: ProfilesController_update
      summary: Обновить данные профиля по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор профиля
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Профиль успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
        '422':
          description: Неверный запрос
      tags: *ref_1
      security: *ref_2
    delete:
      operationId: ProfilesController_remove
      summary: Удалить профиль по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор профиля
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Профиль успешно удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_1
      security: *ref_2
...
  /bots:
    get:
      operationId: BotsController_findMy
      summary: Список ботов пользователя
      parameters: []
      responses:
        '200':
          description: Список ботов пользователя получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: &ref_13
        - bots
      security: &ref_14
        - bearer: []
    post:
      operationId: BotsController_create
      summary: Создание нового бота
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotDto'
      responses:
        '201':
          description: Новый бот создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '422':
          description: Неверный запрос
      tags: *ref_13
      security: *ref_14
  /bots/{id}:
    delete:
      operationId: BotsController_remove
      summary: Удаление бота
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Бот удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_13
      security: *ref_14
    patch:
      operationId: BotsController_update
      summary: Смена имени бота
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                botName:
                  type: string
                  example: Салон красоты
      responses:
        '200':
          description: Имя бота обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_13
      security: *ref_14
    get:
      operationId: BotsController_findOne
      summary: Получить данные бота по Id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Информация о боте по Id получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_13
      security: *ref_14
  /bots/{id}/copy:
    post:
      operationId: BotsController_copy
      summary: Копирование бота
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyBotDto'
      responses:
        '201':
          description: Бот скопирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
        '422':
          description: Неверный запрос
      tags: *ref_13
      security: *ref_14
  /bots/{id}/share:
    post:
      operationId: BotsController_share
      summary: Предоставить общий доступ к боту, первичный доступ при отправке email
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareBotDto'
      responses:
        '201':
          description: Первичный доступ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
        '422':
          description: Неверный запрос
      tags: *ref_13
      security: *ref_14
...
info:
  title: API BotKits
  description: Ручки для команды frontend
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    Credentials:
      type: object
      properties:
        email:
          type: string
          example: my@mail.ru
        password:
          type: string
          example: password123
        accessToken:
          type: string
          example: dkskddksldlslsajsjsdsk,cmksjdksdjskjdk
        refreshToken:
          type: string
          example: dkskddksldlslsajsjsdsk,cmksjdksdjskjdk
      required:
        - email
        - password
        - accessToken
        - refreshToken
    Profile:
      type: object
      properties:
        username:
          type: string
          example: Ivan Ivanov
        phone:
          type: string
          example: '+79501364578'
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        balance:
          type: number
          example: 1400
        accounts:
          type: array
          items:
            type: string
      required:
        - username
        - phone
        - avatar
        - balance
        - accounts
    Account:
      type: object
      properties:
        type:
          type: string
          example: local
        role:
          type: string
          example: admin
        credentials:
          example:
            email: my@mail.ru
            password: password123
            accessToken: dkskddksldlslsajsjsdsk,cmksjdksdjskjdk
            refreshToken: dkskddksldlslsajsjsdsk,cmksjdksdjskjdk
          allOf:
            - $ref: '#/components/schemas/Credentials'
        profile:
          $ref: '#/components/schemas/Profile'
      required:
        - type
        - role
        - credentials
        - profile
    CreateProfileDto:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 30
          example: Ivan Ivanov
        phone:
          type: string
          example: '+79501364578'
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        balance:
          type: number
          example: 1400
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
      required:
        - username
        - phone
    UpdateProfileDto:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 30
          example: Ivan Ivanov
        phone:
          type: string
          example: '+79501364578'
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        balance:
          type: number
          example: 1400
...
    Messenger:
      type: object
      properties:
        name:
          type: string
          example: VK
        page:
          type: string
          example: vk.com/club1245321223
        accessKey:
          type: string
          example: 1685494522:AAHzRs4YFqckLvBVARVoUL0c3B1GFqlDpo
        url:
          type: string
          example: some_url
      required:
        - name
        - page
        - accessKey
        - url
    Bot:
      type: object
      properties:
        icon:
          type: string
          example: >-
            https://cdn.icon-icons.com/icons2/1233/PNG/512/1492718766-vk_83600.png
        botName:
          type: string
          example: Бот Автоответчик
        profile:
          $ref: '#/components/schemas/Profile'
        messenger:
          example:
            name: VK
            page: vk.com/club1245321223
            accessKey: 1685494522:AAHzRs4YFqckLvBVARVoUL0c3B1GFqlDpo
            url: some_url
          allOf:
            - $ref: '#/components/schemas/Messenger'
        botSettings:
          type: object
      required:
        - icon
        - botName
        - profile
        - messenger
        - botSettings
...
    CreateBotDto:
      type: object
      properties:
        icon:
          type: string
          example: >-
            https://cdn.icon-icons.com/icons2/1233/PNG/512/1492718766-vk_83600.png
        botName:
          type: string
          example: Бот Автоответчик
        messenger:
          $ref: '#/components/schemas/Messenger'
        botSettings:
          type: object
          properties: {}
      required:
        - icon
        - botName
        - messenger
        - botSettings
    CopyBotDto:
      type: object
      properties:
        messenger:
          $ref: '#/components/schemas/Messenger'
      required:
        - messenger
    ShareBotDto:
      type: object
      properties:
        email:
          type: string
          description: Email зарегистрированного пользователя
          example: test@test.ru
      required:
        - email
...